+++ src/TableFactory/TableFactory.php
@@ -14,31 +14,33 @@ namespace Phuria\UnderQuery\TableFactory;
 use Phuria\UnderQuery\QueryBuilder\BuilderInterface;
 use Phuria\UnderQuery\Table\SubQueryTable;
 use Phuria\UnderQuery\Table\UnknownTable;
-use Phuria\UnderQuery\TableRecognizer;
-use Phuria\UnderQuery\TableRegistry;
 
 /**
  * @author Beniamin Jonatan Šimko <spam@simko.it>
  */
 class TableFactory implements TableFactoryInterface
 {
-    /**
-     * @var TableRegistry $registry
-     */
-    private $registry;
+    const TYPE_CLOSURE = 1;
+    const TYPE_CLASS_NAME = 2;
+    const TYPE_TABLE_NAME = 3;
+    const TYPE_SUB_QUERY = 4;
 
     /**
-     * @var TableRecognizer $tableRecognizer
+     * @param mixed $stuff
+     *
+     * @return int
      */
-    private $tableRecognizer;
-
-    /**
-     * @param TableRegistry $registry
-     */
-    public function __construct(TableRegistry $registry)
+    public function recognizeType($stuff)
     {
-        $this->registry = $registry;
-        $this->tableRecognizer = new TableRecognizer();
+        if ($stuff instanceof \Closure) {
+            return static::TYPE_CLOSURE;
+        } else if ($stuff instanceof BuilderInterface) {
+            return static::TYPE_SUB_QUERY;
+        } else if (false !== strpos($stuff, '\\')) {
+            return static::TYPE_CLASS_NAME;
+        }
+
+        return static::TYPE_TABLE_NAME;
     }
 
     /**
@@ -46,53 +48,34 @@ class TableFactory implements TableFactoryInterface
      */
     public function createNewTable($table, BuilderInterface $qb)
     {
-        $tableType = $this->tableRecognizer->recognizeType($table);
-
-        $tableClass = null;
+        $tableType = $this->recognizeType($table);
 
         switch ($tableType) {
-            case TableRecognizer::TYPE_CLOSURE:
-                $tableClass = $this->tableRecognizer->extractClassName($table);
-                break;
-            case TableRecognizer::TYPE_CLASS_NAME:
-                $tableClass = $table;
-                break;
-            case TableRecognizer::TYPE_TABLE_NAME:
-                $tableClass = $this->registry->getTableClass($table);
-                break;
-            case TableRecognizer::TYPE_SUB_QUERY:
-                return $this->createSubQueryTable($table, $qb);
+            case static::TYPE_CLOSURE:
+                $tableClass = $this->extractClassName($table);
+                return new $tableClass($qb);
+            case static::TYPE_CLASS_NAME:
+                return new $table($qb);
+            case static::TYPE_SUB_QUERY:
+                return new SubQueryTable($table, $qb);
         }
 
-        return $this->doCreate($table, $tableClass, $qb);
-    }
-
-    /**
-     * @param string           $requestedTable
-     * @param string           $tableClass
-     * @param BuilderInterface $qb
-     *
-     * @return mixed
-     */
-    private function doCreate($requestedTable, $tableClass, BuilderInterface $qb)
-    {
-        $tableObject = new $tableClass($qb);
-
-        if ($tableObject instanceof UnknownTable) {
-            $tableObject->setTableName($requestedTable);
-        }
+        $tableObject = new UnknownTable($qb);
+        $tableObject->setTableName($table);
 
         return $tableObject;
     }
 
     /**
-     * @param BuilderInterface $subQb
-     * @param BuilderInterface $qb
+     * @param \Closure $closure
      *
-     * @return SubQueryTable
+     * @return string
      */
-    private function createSubQueryTable(BuilderInterface $subQb, BuilderInterface $qb)
+    public function extractClassName(\Closure $closure)
     {
-        return new SubQueryTable($subQb, $qb);
+        $ref = new \ReflectionFunction($closure);
+        $firstParameter = $ref->getParameters()[0];
+
+        return $firstParameter->getClass()->getName();
     }
 }