+++ CHANGELOG.md
@@ -4,7 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-0.7.3 [2016-11-18]
+Unreleased
+---------
+
+### Added
+
+### Changed
++ TableRecognizer logic moved to TableFactory
+
+### Removed
++ Container, DependencyInjection and pimple/pimple package
++ TableRegistry
+
+0.7.4 [2016-11-18]
 ----------------
 
 ### Added